refactor components directory
