+++ client/src/components/AuthWidget.js
