+++ client/src/components/AvailabilityWidget.js
