+++ client/src/components/EventBodyComponent/EventBodyComponent.js
