+++ client/src/components/SecondAuthWidget.js
